// part9/patientor-backend/types.ts
export interface Diagnosis {
  code: string;
  name: string;
  latin?: string;
}

export enum Gender {
  Male = "male",
  Female = "female",
  Other = "other"
}

export interface Patient {
  id: string;
  name: string;
  occupation: string;
  gender: string;
  ssn?: string;
  dateOfBirth?: string;
}

export type PatientNoSSN = Omit<Patient, 'ssn'>;

export type PatientFormValues = Omit<Patient, "id">;